@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   connection: {
     protocol: 'amqp',
     name: 'default',
