@@ -1,8 +1,11 @@
-const chai = require('chai');
-chai.use(require('chai-as-promised'));
+import chai from 'chai';
+import chaiAsPromised from 'chai-as-promised';
+import sinon from 'sinon';
+
+chai.use(chaiAsPromised);
 global.should = chai.should();
 global.expect = chai.expect;
-global.sinon = require('sinon');
+global.sinon = sinon;
 process.title = 'rabbot-test';
 
 global.harnessFactory = function (rabbit, cb, expected) {
@@ -66,12 +69,12 @@ global.harnessFactory = function (rabbit, cb, expected) {
       received.push(msg);
       check();
     },
-    received: received,
-    clean: clean,
+    received,
+    clean,
     handle: handleFn,
-    handlers: handlers,
-    unhandled: unhandled,
-    returned: returned
+    handlers,
+    unhandled,
+    returned
   };
 };
 
